There was a very minor error with argv in HLiam, but this revealed a much bigger mistake on my part.
I hadn't thought of what would happen when other people would send me patches, and how they would
figure out my baroque testing regime.

So I've updated my Makefile, testing stuff and so on, so that it's a bit more foolproof. Once the
git hooks are in place, you shouldn't be able to commit untested things. Furthermore, if the
git hooks aren't in place, I should get some sort of indication that this is happening, via the
commit message.
commit-msg approval for sebastienloisel on 2020.03.08-21:23:03