Changes this patch were complicated but user-facing things, not so much. The major changes all have to do with matplotlib.
I now briefly describe what I learned about matplotlib. It seems that matplotlib keeps a list
of figures that are supposed to be on-screen at the moment. When you close those figures, the
corresponding Python data structures get deleted or otherwised closed. Each Figure is assigned
a FigureManager, and when the Figure is closed at the GUI level, the corresponding FigureManager
is deleted.

In a normal Python program, each Figure object is displayed in a separate window. As the Python
program runs, the Figure may be modified, and eventually, this Figure window is closed. All in
all, the Figure is almost tangential to the Python program.

In document generation, the Figure must appear at a specific place in the document. In a way,
this is similar to what Jupyter does. I must confess I do not completely understand how
Jupyter handles matplotlib plots. Long story short, I became worried that I would write
to an undocumented API that was going to change between minor versions of matplotlib.
This would have been bad.

So I decided to completely avoid the whole GUI infrastructure in matplotlib, in the
following ways. First, the various show() functions now do nothing. In standard
matplotlib, these functions either cause the Figure to pop up on screen, or otherwise
cause a fatal error in the Python interpreter, if that is impossible. Neither of these
outcomes are desirable in PypTeX.

Second, I've made it so that Figure objects, when pretty-printed, cause a relevant
.eps file to be written out, and the appropriate \includegraphics command to be
emitted to the relevant .tex file. So, if fig is a Figure object, pyp.print(fig)
or a corresponding pyp.pp(...) call, will cause the fig object to be saved to a file,
and emit the appropriate \includegraphics command.

Third, I've made it so that any unprinted Figure is automatically emitted at the
end of any Python block. I'm now going to explain how this works, and it will sound
super complicated, but I think this is how Jupyter does it, and I think it's not so
bad.

The short version is that, if a Figure has never been printed, then it will be
printed at the end of the current Python block of code. Any time PypTeX prints
a Figure X, it sets X.drawn=True. The point of this is that,
if X.drawn==True, then that Figure does not get printed at the end of the Python block.

You can set pyp.autoshow=False if you want to suppress this behavior.

In addition to all of this, I've made the following minor modification. You can
now use the notation "{some python expression}" in pyptex.pp format strings.
Previously, you could do things like pyp.pp("x is @x"), but now you can do
pyp.pp("x is @{x+7}")

This won't work so great inside of @{...} blocks in your .tex source file,
but it will be fine inside of @{{{...}}}

commit-msg approval for sebastienloisel on 2022.02.11-22:43:36