Improved error reporting and pdb behavior.
When an Exception is generated by user code in a file a.tex that
is being compiled by PypTeX, the underlying __traceback__ is
truncated so that the topmost stack frame is that of a.tex,
consistent with the illusion that a.tex is somehow the topmost
scope that is running in Python.

If somehow the PypTeX compiler itself were to raise an Exception
(this would be a bug in the PypTeX compiler), those stacks are
not truncated and will top out in the PypTeX module itself, or
wherever the Exception is eventually caught.

commit-msg approval for sebastienloisel on 2023.01.11-16:09:16