Minor tweaks to Exception handling in user code.
Instead of calling a "postmortem" function inside
the exec_and_catch function, I now raise an
ExecError Exception in exec_and_catch, which can
be caught and handled by the caller. This felt
slightly more Pythonic, but does not change the
experience of command-line PypTeX users.

commit-msg approval for sebastienloisel on 2023.01.11-17:06:53