Fixes related to traceback truncation within pyp.pp(...)
PypTeX iterates through Python fragment in the input file
a.tex and runs each such fragment using either exec or eval.
I shall refer to those as the "outer" calls to exec and eval.
Within such fragment, it is common to use pyp.pp(...) to
pretty-print various expressions into the ouput a.pyptex file.
The function pyp.pp(...) makes "inner" calls to eval.
If these inner calls to eval generate Exceptions, the traceback
should not be truncated, because it should clearly go all the
way up to a.tex

More generally, if a Python fragment in a.tex embeds another
file, e.g. by doing pyptex('b.tex'), it is preferable
not to truncate the top of the stack in b.tex, so that it
may be followed all the way back up to a.tex. It is somewhat
unfortunate that this will go through a bunch of layers of
PypTeX itself, but I'm not sure what to do about it.

commit-msg approval for sebastienloisel on 2023.01.11-17:36:19