Restore stack filtering that had been accidentally removed
When there is a Python exception, the stack trace is supposed to
be filtered so that internal pyptex stack frames are deleted.
I had accidentally suppressed that behavior in the previous
version.

There are some slight changes elsewhere that are related to my
trying to track down some sort of race condition or python bug
or even possibly an OS bug. PypTeX uses streamcapture to
capture the output of spawned processes and log them. This
machinery is now also used to annote LaTeX error messages
with line numbers in the intermediary .pyptex file, with the
corresponding line numbers in the source .tex file, for easy
cross-referencing.

I'm not sure exactly what happened, but apparently the slightly
increased workload in the streamcapture "writer", which detects
those LaTeX error messages, triggered some sort of race condition
or underlying python or OS bug. It's hard to debug this because,
by its nature, streamcapture redirects stdout and stderr, so
as soon as streamcapture errors out, you often don't have stdout
and stderr anymore, so pdb is not so useful. Instead, I used
webpdb, which at least allows me to debug my program a bit. The
second problem is that streamcapture is inherently threaded, so
you have to launch webpdb in the right thread or else you won't
be able to debug your problem.

Long story short, from the master thread, it looks like sometimes,
when you os.close() the write end of a pipe, it doesn't actually
close it. The problem with this is that the streamcapture thread
that reads from the pipe, never gets an EOF, so that thread becomes
a zombie, preventing the whole process from terminating. One
workaround would be to daemonize the slave thread(s) but that is
also wrong, because the slave threads need to empty whatever
data is pending from the pipe and write it to whatever log file,
before finishing up, and daemonization would prevent that.

The weird thing is that, supposedly, the underlying OS should
signal an error when the underlying C library close() function
fails, and errno should be said, and according to python documentation,
this should trigger an exception. I never saw any of this.
It's not immediately obvious how to read libc's errno directly
from Python (I guess maybe ctypes or something).

Reading on the internet people struggling with pipes in C,
many people do a complicated song and dance to try and close
these pipes correctly, and often I feel like that code is wrong.
I also read reports of underlying OS bugs with the closing of
those pipes.

Since that seemed like a wild goose chase for now, I implemented
a kludge in streamcapture. When a streamcapture is closed,
it sends a magic byte string through the pipe before closing it.
The receiving thread will strip that magic byte string and cleanup
and shut down. Seems to work for now.

commit-msg approval for sebastienloisel on 2023.01.27-11:49:31